feat:10连抽接口完成,测试通过

+++ big-market-domain/src/main/java/com/openicu/domain/strategy/resposity/IStrategyRepository.java
@@ -217,7 +217,7 @@ public interface IStrategyRepository {
      * @param strategyId
      * @return
      */
-    Integer queryActivityAccountTotalUseCount(String userId, Long strategyId);
+    Long queryActivityAccountTotalUseCount(String userId, Long strategyId);
 
     /**
      * 查询运行的抽奖策略奖品列表
